requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-resized==1.0.2
django-shortcuts==1.6
djangorestframework==3.14.0
pillow==10.2.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
